ref(ui): added types.tsx
